# docs/openapi.yaml
openapi: 3.0.3
info:
  title: ARCode Clone API
  description: API REST complète pour l'application ARCode Clone iOS
  version: 1.0.0
  contact:
    name: ARCode Support
    email: support@ar-code.com
    url: https://ar-code.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.ar-code.com
    description: Production
  - url: https://staging-api.ar-code.com
    description: Staging
  - url: http://localhost:8080
    description: Local Development
...
paths:
...
  /api/v1/ab-tests/{ar_code_id}:
    get:
      tags:
        - A/B Testing
      summary: Get AB Test
      description: Récupérer le test A/B actif d'un AR Code
      parameters:
        - name: ar_code_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: AB Test or null
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ABTest'
  /api/v1/ab-tests:
    post:
      tags:
        - A/B Testing
      summary: Create AB Test
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ABTestCreate'
      responses:
        '201':
          description: AB Test created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ABTest'
  /api/v1/ab-tests/{test_id}/results:
    get:
      tags:
        - A/B Testing
      summary: Get AB Test Results
      parameters:
        - name: test_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: AB Test results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ABTestResults'
  /api/v1/workspaces:
    get:
      tags:
        - Workspaces
      summary: Get Workspaces
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of workspaces
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Workspace'

    post:
      tags:
        - Workspaces
      summary: Create Workspace
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                description:
                  type: string
      responses:
        '201':
          description: Workspace created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workspace'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ABTest:
      type: object
      properties:
        id:
          type: string
          format: uuid
        ar_code_id:
          type: string
          format: uuid
        name:
          type: string
        is_active:
          type: boolean
        variants:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              weight:
                type: integer
              cta_link_id:
                type: string
                format: uuid
        start_date:
          type: string
          format: date-time
        end_date:
          type: string
          format: date-time
          nullable: true
        winner_variant_id:
          type: string
          nullable: true
    ABTestCreate:
      type: object
      required:
        - ar_code_id
        - name
        - variants
      properties:
        ar_code_id:
          type: string
          format: uuid
        name:
          type: string
        variants:
          type: array
          items:
            type: object
        is_active:
          type: boolean
          default: true
    ABTestResults:
      type: object
      properties:
        test_id:
          type: string
          format: uuid
        variants:
          type: array
          items:
            type: object
            properties:
              variant_id:
                type: string
              clicks:
                  type: integer
              conversions:
                  type: integer
              conversion_rate:
                  type: number
    Workspace:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
          nullable: true
        owner_id:
          type: string
          format: uuid
        settings:
          type: object
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...